+++ kok-api/src/main/java/com/kok/kokapi/common/adapter/in/web/BaseController.java
@@ -1,4 +1,4 @@
-package com.kok.kokapi.adapter.in.web;
+package com.kok.kokapi.common.adapter.in.web;
 
 import org.springframework.web.bind.annotation.RequestMapping;
